+++ test/aohara/tinkertime/UnitTestSuite.java
@@ -1,23 +1,23 @@
-package aohara.tinkertime.testSuites;
+package aohara.tinkertime;
 
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
-import aohara.tinkertime.TestModStateManager;
-import aohara.tinkertime.TestModStructure;
-import aohara.tinkertime.TestModuleManagerCrawler;
 import aohara.tinkertime.crawlers.TestCrawlerFactory;
 import aohara.tinkertime.crawlers.TestCurseCrawler;
 import aohara.tinkertime.crawlers.TestGithubHtmlCrawler;
 import aohara.tinkertime.crawlers.TestGithubJsonCrawler;
 import aohara.tinkertime.crawlers.TestKerbalStuffCrawler;
+import aohara.tinkertime.crawlers.TestJenkinsCrawler;
+import aohara.tinkertime.resources.TestModLoader;
+import aohara.tinkertime.resources.TestModStructure;
 
 
 @RunWith(Suite.class)
 @Suite.SuiteClasses({
    TestCurseCrawler.class,
-   TestModStateManager.class,
-   TestModuleManagerCrawler.class,
+   TestModLoader.class,
+   TestJenkinsCrawler.class,
    TestGithubHtmlCrawler.class,
    TestGithubJsonCrawler.class,
    TestCrawlerFactory.class,
